Moved createKeyValueLedger into package object.
